Move non GPX elements from PlotGPX to PlotTracks
